requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.41
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
websockets==12.0
requests==2.31.0
confluent-kafka>=2.6.0
paho-mqtt==1.6.1
databricks-sql-connector>=3.5.0
